refactored getAccountData to separete file
